Fix pre-existing user being unable to check in
If a user's email in the system has any uppercase letters in it, it will
not be found when that person tries to check in. This caused a bug where
a user's email was not being found as a pre-existing user, so they were
being asked to sign in anew--but then the system was rejecting the
signup because the email was already in the system. They could not
proceed.